+++ Sources/Core/AsynchronousOperation.swift
@@ -3,23 +3,6 @@
 //
 // Copyright © 2016-2020 Tinrobots.
 //
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
 
 import Foundation
 
@@ -30,7 +13,7 @@ public typealias AsyncOperation = AsynchronousOperation
 /// Subclasses must override `main` to perform any work and, if they are asynchronous, call the `finish()` method to complete the execution.
 open class AsynchronousOperation: Operation, ProgressReporting {
   // MARK: - Public Properties
-  
+
   /// The `progress` property represents a total progress of the operation during its execution.
   @objc
   public final lazy private(set) var progress: Progress = {
@@ -39,40 +22,33 @@ open class AsynchronousOperation: Operation, ProgressReporting {
     progress.isCancellable = true
     progress.cancellationHandler = { [weak self] in
       if let self = self, !self.isCancelled {
-        print(progress.isCancelled)
         self.cancel()
       }
     }
     return progress
   }()
-  
-  open override var isReady: Bool {
-    return state == .ready && super.isReady
-  }
-  
-  public final override var isExecuting: Bool {
-    return state == .executing
-  }
-  
-  public final override var isFinished: Bool {
-    return state == .finished
-  }
-  
-  public final override var isAsynchronous: Bool { return isConcurrent }
-  
-  public final override var isConcurrent: Bool { return true }
-  
+
+  open override var isReady: Bool { state == .ready && super.isReady }
+
+  public final override var isExecuting: Bool { state == .executing }
+
+  public final override var isFinished: Bool { state == .finished }
+
+  public final override var isAsynchronous: Bool { isConcurrent }
+
+  public final override var isConcurrent: Bool { true }
+
   // MARK: - Private Properties
-  
+
   /// Lock used to prevent data races when updating the progress.
   private let progressLock = UnfairLock()
-  
+
   /// Serial queue for making state changes atomic under the constraint of having to send KVO willChange/didChange notifications.
   private let stateChangeQueue = DispatchQueue(label: "\(identifier).AsynchronousOperation.stateChange")
-  
+
   /// Private backing store for `state`
   private var _state: Atomic<State> = Atomic(.ready)
-  
+
   /// The state of the operation
   private var state: State {
     get {
@@ -93,30 +69,30 @@ open class AsynchronousOperation: Operation, ProgressReporting {
         // willChange/didChange notifications only for the key paths that actually change.
         let oldValue = _state.value
         guard newValue != oldValue else { return }
-        
+
         willChangeValue(forKey: newValue.objcKeyPath)
         willChangeValue(forKey: oldValue.objcKeyPath)
-        
+
         _state.mutate {
           assert($0.canTransition(to: newValue), "Performing an invalid state transition from: \($0) to: \(newValue) for \(operationName).")
           $0 = newValue
         }
-        
+
         didChangeValue(forKey: oldValue.objcKeyPath)
         didChangeValue(forKey: newValue.objcKeyPath)
       }
     }
   }
-  
+
   // MARK: - Foundation.Operation
-  
+
   // Lock used to prevent data races if start() gets called multiple times from different threads
   private let startLock = UnfairLock()
-  
+
   public final override func start() {
     startLock.lock()
     defer { startLock.unlock() }
-    
+
     switch state {
     case .finished:
       return
@@ -127,15 +103,15 @@ open class AsynchronousOperation: Operation, ProgressReporting {
       guard isReady else {
         fatalError("The operation \(operationName) is not yet ready to execute.")
       }
-      
+
       // early bailing out
       guard !isCancelled else {
         finish()
         return
       }
-      
+
       state = .executing
-      
+
       // The OperationQueue progress reporting works correcly only if used with super.start()
       // but calling super.start() shouldn't be done when implementing custom concurrent operations.
       // To fix that we use a different progress instance
@@ -146,15 +122,16 @@ open class AsynchronousOperation: Operation, ProgressReporting {
         }
         progressLock.unlock()
       }
+
       main()
-      
+
       // At this point `main()` has already returned but it doesn't mean that the operation is finished.
       // Only calling `finish()` will finish the operation at this point.
     }
   }
-  
+
   // MARK: - Public Methods
-  
+
   ///  The default implementation of this method does nothing.
   /// You should override this method to perform the desired task. In your implementation, do not invoke super.
   ///  This method will automatically execute within an autorelease pool provided by Operation, so you do not need to create your own autorelease pool block in your implementation.
@@ -162,10 +139,10 @@ open class AsynchronousOperation: Operation, ProgressReporting {
   open override func main() {
     preconditionFailure("Subclasses must implement `main()`.")
   }
-  
+
   /// Finishes the operation.
-  /// - Important: You should never call this method outside the operation main execution scope.
-  public final func finish() {
+  /// - Important: You should never call this method outside the operation main() execution scope.
+  public func finish() {
     // State can also be "ready" here if the operation was cancelled before it was started.
     if !isFinished {
       progressLock.lock()
@@ -173,7 +150,7 @@ open class AsynchronousOperation: Operation, ProgressReporting {
         progress.completedUnitCount = progress.totalUnitCount
       }
       progressLock.unlock()
-      
+
       // If multiple calls are made to finish() from different threads at the same time
       // setting the same state will trigger an assert in the state setter.
       // A lock isn't required.
@@ -182,25 +159,23 @@ open class AsynchronousOperation: Operation, ProgressReporting {
       preconditionFailure("The finish() method shouldn't be called more than once for \(operationName).")
     }
   }
-  
+
   open override func cancel() {
     super.cancel()
-    
+
     progressLock.lock()
     if !progress.isCancelled {
       progress.cancel()
     }
     progressLock.unlock()
   }
-  
+
   // MARK: - Debug
-  
-  open override var description: String {
-    return debugDescription
-  }
-  
+
+  open override var description: String { debugDescription }
+
   open override var debugDescription: String {
-    return "\(operationName) – \(isCancelled ? "cancelled (\(state))" : "\(state)")"
+    "\(operationName) – \(isCancelled ? "cancelled (\(state))" : "\(state)")"
   }
 }
 
@@ -212,7 +187,7 @@ extension AsynchronousOperation {
     case ready // waiting to be executed
     case executing
     case finished
-    
+
     /// The `#keyPath` for the `Operation` property that's associated with this value.
     var objcKeyPath: String {
       switch self {
@@ -221,7 +196,7 @@ extension AsynchronousOperation {
       case .finished: return #keyPath(isFinished)
       }
     }
-    
+
     var description: String {
       switch self {
       case .ready: return "ready"
@@ -229,11 +204,11 @@ extension AsynchronousOperation {
       case .finished: return "finished"
       }
     }
-    
+
     var debugDescription: String {
       return description
     }
-    
+
     func canTransition(to newState: State) -> Bool {
       switch (self, newState) {
       case (.ready, .executing): return true