+++ Sources/Core/Operation+Utils.swift
@@ -3,36 +3,15 @@
 //
 // Copyright © 2016-2020 Tinrobots.
 //
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
 
 import Foundation
 
 extension Operation {
   /// Returns the `Operation` name or its type if the name is nil.
-  public var operationName: String {
-    return name ?? "\(type(of: self))"
-  }
+  public var operationName: String { name ?? "\(type(of: self))" }
 
   /// Returns `true` if **at least one** dependency has been cancelled.
-  public var hasSomeCancelledDependencies: Bool {
-    return dependencies.hasSomeCancelledOperations
-  }
+  public var hasSomeCancelledDependencies: Bool { dependencies.hasSomeCancelledOperations }
 
   /// Adds multiple dependencies to the operation.
   /// If the receiver is already executing its task, adding dependencies has no practical effect.
@@ -79,7 +58,5 @@ extension OperationQueue {
 
 extension Collection where Element: Operation {
   /// Returns `true` if **at least one** operation has been cancelled.
-  fileprivate var hasSomeCancelledOperations: Bool {
-    return first { $0.isCancelled } != nil
-  }
+  fileprivate var hasSomeCancelledOperations: Bool { first { $0.isCancelled } != nil }
 }